Update index.html
a few minor revisions.

Our proposed distributed algorithm for opinion updates is also related to the so-called block-coordinate gradient descent/ascent (with/without projection) algorithm and alternating optimization algorithms [2] in continuous optimization. If an agent’s opinion is seen as a block of variables, our algorithm can distributedly update each agent’s opinion in parallel compared to the sequential updates of blocks, a limitation in some sense, using gradient descent/ascent or direct optimization. Gradient updates tend to help stabilize the dynamics compared to the abrupt update of one-step optimization, which is deemed the best response.